Move Delayed::Job plugin to integrations namespace
The DelayedJobPlugin was defined in the integrations directory, but used
the Hooks module namespace. Update the namespace so it matches its
location. No redirect needed because it's a private API no one should be
accessing.

Split the test file so it tests what's in the related file.
